requirements.txt
streamlit==1.31.0
tensorflow-cpu==2.15.0
numpy
pillow
